# StoriesService/static/api.yaml
swagger: '2.0'
info:
  title: Stories Service
  description: Avalaible operation for stories
  version: '0.1'
host: 127.0.0.1
schemes:
  - http
  - https
paths:
  /stories:
    get:
      summary: Returns a list of all the stories
      operationId: getStories
      produces:
        - application/json
      responses:
        '200':
          description: Array of story as described in definitions
          schema:
            type: array
            items:
              $ref: '#/definitions/story'
  /stories/{id_story}:
    get:
      summary: Return the story specified by id_story
      operationId: getStory
      parameters:
        - in: path
          name: id_story
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '404':
          description: Specified story not found
        '200':
          description: Object story descibed in definitions
          schema:
            $ref: '#/definitions/story'
  /stories/latest:
    get:
      summary: Gets the last NON-draft story for each registered user
      operationId: getLatestStories
      produces:
        - application/json
      responses:
        '200':
          description: Array of story as described in definitions
          schema:
            type: array
            items:
              $ref: '#/definitions/story'
  /stories/range:
    get:
      summary: Searches for stories that were made in a specific range of time
      operationId: getRangeStories
      parameters:
        - in: query
          name: begin
          description: Range begin in 'yyyy-mm-dd' format
          type: string
        - in: query
          name: end
          description: Range end in 'yyyy-mm-dd' format
          type: string
      produces:
        - application/json
      responses:
        '400':
          description: Wrong URL parameters/Begin date cannot be higher than End date
        '200':
          description: Array of story as described in definitions
          schema:
            type: array
            items:
              $ref: '#/definitions/story'
  /stories/random:
    get:
      summary: Get a random story written by other users in the last three days
      operationId: getRandomStory
      parameters:
        - in: query
          name: user_id
          description: Current user id
          type: integer
      produces:
        - application/json
      responses:
        '404':
          description: There are no recent stories by other users
        '200':
          description: Object story descibed in definitions
          schema:
            $ref: '#/definitions/story'
  /stories/new/write/{id_story}:
    get:
      summary: Set the session to contine a draft
      operationId: writeDraft
      parameters:
        - in: path
          name: id_story
          description: id of draft you want continue
          required: true
          type: integer
        - in: query
          name: user_id
          description: Current user id
          type: integer
      responses:
        '400':
          description: Request is invalid, check if you are the author of the story and it is still a draft
        '200':
          description: Session  to continue writing a draft OK
  /stories/new/write:
    post:
      summary: Submit a story or a draft
      operationId: writeStory
      consumes:
        - application/json
      parameters:
        - in: body
          name: story_submit
          schema:
            $ref: '#/definitions/story_submit'
      responses:
        '400':
          description: Errors in requestbody
        '422':
          description: Story doesn't contain all the words or it is too long
        '201':
          description: Draft created / Draft updated / Draft has been published / Story has been published
  /stories/stats/{user_id}:
    get:
      summary: Get the statistics for a specific user
      operationId: getStoriesStatistics
      consumes:
        - application/json
      parameters:
        - in: path
          name: user_id
          description: The user id
          required: true
          type: integer
      responses:
        '400':
          description: Errors in requestbody
        '200':
          description: Statistics of stories for a given user
  /stories/delete/{id_story}:
    post:
      summary: Delete the story specified by story id
      operationId: deleteStory
      parameters:
        - in: path
          name: id_story
          description: Id of story you want delete
          required: true
          type: integer
        - in: query
          name: user_id
          description: Current user id
          type: integer
      responses:
        '400':
          description: Request is invalid, check if you are the author of the story and the id is a valid one
        '200':
          description: Story has been deleted
  /search:
    get:
      summary: Return the list of all matching stories
      operationId: search
      produces:
        - application/json
      responses:
        '200':
          description: A JSON array of JSON objects containing stories list
          schema: 
            type: array
            items: 
              $ref: '#/definitions/story'
        '400':
          description: Error with query parameter
definitions:
  story:
    type: object
    properties:
      id:
        type: integer
        description: Story id
      text:
        type: string
        description: Story text
      date:
        type: string
        description: DateTime of submission
      figures:
        type: string
        description: Mandatory words in the story
      author_id:
        type: integer
        description: Author's id
      is_draft:
        type: boolean
        description: True if it is a draft, false if it has been published
  story_submit:
    type: object
    properties:
      text:
        type: string
        description: Story text to submit
      as_draft:
        type: boolean
        description: True to save a draft, false otherwise
      user_id:
        type: integer
        description: the author of the story
        